requirements.txt
numpy
scikit-image
scipy
pillow<=6.2.2
torchvision
tensorboardX
